Add theme exports to hooks and components
