+++ frontend/components/utils/theme/hueMap.ts
