+++ frontend/hooks/contexts/UserContext.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { createContext } from 'react';
 import { type FetchedUser } from 'codetierlist-types';
-import { defaultAccentColor } from '@/utils/theme';
+import { defaultAccentColor } from '@/components/utils/theme/theme';
 
 /**
  * The user that is shown when the user's information is still loading.
