+++ frontend/hooks/index.ts
@@ -4,3 +4,5 @@ export * from './contexts/UserContext';
 export * from './custom/useMediaQuery';
 export * from './custom/useSeed';
 export * from './custom/useSystemTheme';
+
+export * from './theme/useTheme';
